feature: updates Commandline project to execute prompts via calling Api instead of directly, lowering duplicate code
Project can technically be substituted for a service online that simple pings the correct endpoint on a desired timer.
The Project will however still remain useful for use to generate migrations, when entities are changed, as it remains configured to setup a connection to the database - just with zero use of it.